fix(table): export full tables as images via slice capture fallback
Naive toImage at 3x silently clipped tables taller than the GPU max texture size (issue #252). Extract the message-export capture machinery (ratio cap, clip detection, overlay slice capture, isolate composite, optional blank trim) into lib/utils/export_image_capture.dart and apply it to the markdown table path with no trim. The slice rebuild reuses the in-tree table surface under its captured MediaQuery so desktop chat font scale stays pixel-identical.
