+++ lib/features/chat/widgets/message_export_sheet.dart
@@ -1,12 +1,9 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
-import 'dart:math' as math;
-import 'dart:ui' as ui;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:image/image.dart' as image_lib;
 import 'package:intl/intl.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/rendering.dart';
@@ -25,6 +22,7 @@ import '../../../core/providers/user_provider.dart';
 import '../../../core/providers/assistant_provider.dart';
 import '../../../core/models/assistant.dart';
 import '../../../core/services/chat/chat_service.dart';
+import '../../../utils/export_image_capture.dart';
 import '../../../utils/sandbox_path_resolver.dart';
 import '../../../shared/widgets/markdown_with_highlight.dart';
 import '../../../shared/widgets/export_capture_scope.dart';
@@ -777,9 +775,6 @@ Future<File?> _renderAndSaveChatImage(
 const double _desktopExportLogicalWidth = 720.0;
 const double _mobileExportLogicalWidth = 480.0;
 const double _exportImagePixelRatio = 3.0;
-const int _exportImageBlankTrimPreservePaddingPhysical = 48;
-const int _exportImageBlankAlphaTolerance = 8;
-const int _exportImageBlankColorTolerance = 3;
 
 ({double width, double pixelRatio}) _exportImageRenderConfig({
   required bool isDesktop,
@@ -832,7 +827,7 @@ Future<File?> _renderWidgetDirectly(
       return Positioned(
         left: -10000, // Position far offscreen
         top: -10000,
-        child: _ExportCaptureRoot(
+        child: ExportCaptureRoot(
           theme: theme,
           boundaryKey: boundaryKey,
           width: width,
@@ -857,18 +852,23 @@ Future<File?> _renderWidgetDirectly(
     if (boundary == null) return null;
 
     final contentSize = boundary.size;
-    var data = await _captureBoundaryPngBytes(boundary, pixelRatio: pixelRatio);
+    var data = await captureBoundaryPngBytes(
+      boundary,
+      pixelRatio: pixelRatio,
+      preservePadding: exportImageBlankTrimPreservePaddingPhysical,
+    );
     if (data == null && !contentSize.isEmpty) {
       entry.remove();
       entryRemoved = true;
       await WidgetsBinding.instance.endOfFrame;
-      data = await _captureWidgetViewportSlicesPngBytes(
+      data = await captureWidgetViewportSlicesPngBytes(
         overlay,
         buildContent,
         theme: theme,
         width: width,
         pixelRatio: pixelRatio,
         contentSize: contentSize,
+        preservePadding: exportImageBlankTrimPreservePaddingPhysical,
       );
     }
     if (data == null) return null;
@@ -884,643 +884,6 @@ Future<File?> _renderWidgetDirectly(
   } finally {
     if (!entryRemoved) entry.remove();
   }
-}
-
-@visibleForTesting
-Widget buildExportCaptureRootForTesting({
-  required ThemeData theme,
-  required Widget child,
-  double width = 480,
-}) {
-  return _ExportCaptureRoot(
-    theme: theme,
-    boundaryKey: GlobalKey(),
-    width: width,
-    child: child,
-  );
-}
-
-@visibleForTesting
-Widget buildExportCaptureViewportRootForTesting({
-  required ThemeData theme,
-  required Widget child,
-  required double viewportHeight,
-  required double contentHeight,
-  double width = 480,
-  double offsetY = 0,
-}) {
-  return _ExportCaptureViewportRoot(
-    theme: theme,
-    boundaryKey: GlobalKey(),
-    width: width,
-    viewportHeight: viewportHeight,
-    contentHeight: contentHeight,
-    offsetY: offsetY,
-    child: child,
-  );
-}
-
-class _ExportCaptureRoot extends StatelessWidget {
-  const _ExportCaptureRoot({
-    required this.theme,
-    required this.boundaryKey,
-    required this.width,
-    required this.child,
-  });
-
-  final ThemeData theme;
-  final GlobalKey boundaryKey;
-  final double width;
-  final Widget child;
-
-  @override
-  Widget build(BuildContext context) {
-    return Theme(
-      data: theme,
-      child: RepaintBoundary(
-        key: boundaryKey,
-        child: Container(
-          width: width,
-          color: theme.colorScheme.surface,
-          child: Material(type: MaterialType.transparency, child: child),
-        ),
-      ),
-    );
-  }
-}
-
-class _ExportCaptureViewportRoot extends StatelessWidget {
-  const _ExportCaptureViewportRoot({
-    required this.theme,
-    required this.boundaryKey,
-    required this.width,
-    required this.viewportHeight,
-    required this.contentHeight,
-    required this.offsetY,
-    required this.child,
-  });
-
-  final ThemeData theme;
-  final GlobalKey boundaryKey;
-  final double width;
-  final double viewportHeight;
-  final double contentHeight;
-  final double offsetY;
-  final Widget child;
-
-  @override
-  Widget build(BuildContext context) {
-    return Theme(
-      data: theme,
-      child: RepaintBoundary(
-        key: boundaryKey,
-        child: Container(
-          width: width,
-          height: viewportHeight,
-          color: theme.colorScheme.surface,
-          child: Material(
-            type: MaterialType.transparency,
-            child: ClipRect(
-              child: OverflowBox(
-                alignment: Alignment.topCenter,
-                minWidth: width,
-                maxWidth: width,
-                minHeight: contentHeight,
-                maxHeight: contentHeight,
-                child: Transform.translate(
-                  offset: Offset(0, -offsetY),
-                  child: child,
-                ),
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-// Keep whole-image captures below the common 16384px GPU texture edge while
-// avoiding the slice compositor for exports that can still fit at >=2x.
-const double _maxExportFullCapturePhysicalDimension = 15360.0;
-const double _maxExportCaptureSlicePhysicalHeight = 4096.0;
-const double _minExportFullCapturePixelRatio = 2.0;
-
-@visibleForTesting
-Future<Uint8List?> captureExportBoundaryPngBytesForTesting(
-  RenderRepaintBoundary boundary, {
-  required double pixelRatio,
-}) {
-  return _captureBoundaryPngBytes(boundary, pixelRatio: pixelRatio);
-}
-
-@visibleForTesting
-bool shouldUseFullExportCaptureForTesting({
-  required Size logicalSize,
-  required double pixelRatio,
-}) {
-  return _shouldUseFullExportCapture(logicalSize, pixelRatio: pixelRatio);
-}
-
-@visibleForTesting
-double? exportFullCapturePixelRatioForTesting({
-  required Size logicalSize,
-  required double requestedPixelRatio,
-}) {
-  return _exportFullCapturePixelRatio(
-    logicalSize,
-    requestedPixelRatio: requestedPixelRatio,
-  );
-}
-
-@visibleForTesting
-double exportCaptureSliceLogicalHeightForTesting({required double pixelRatio}) {
-  return _exportCaptureSliceLogicalHeight(pixelRatio: pixelRatio);
-}
-
-@visibleForTesting
-Uint8List stitchExportPngSlicesForTesting({
-  required int outputWidth,
-  required int outputHeight,
-  required List<({Uint8List bytes, int y})> slices,
-}) {
-  return _stitchExportPngSlices(
-    outputWidth: outputWidth,
-    outputHeight: outputHeight,
-    slices: slices,
-  );
-}
-
-Future<Uint8List?> _captureBoundaryPngBytes(
-  RenderRepaintBoundary boundary, {
-  required double pixelRatio,
-}) async {
-  if (boundary.size.isEmpty) return null;
-  final fullCapturePixelRatio = _exportFullCapturePixelRatio(
-    boundary.size,
-    requestedPixelRatio: pixelRatio,
-  );
-  if (fullCapturePixelRatio != null) {
-    final fullCapture = await _captureFullBoundaryPngBytes(
-      boundary,
-      pixelRatio: fullCapturePixelRatio,
-    );
-    if (fullCapture != null) {
-      return _processCapturedExportPng(
-        _CapturedExportPngProcessingRequest.single(
-          singlePngBytes: fullCapture,
-          preservePadding: _exportImageBlankTrimPreservePaddingPhysical,
-        ),
-      );
-    }
-  }
-  return null;
-}
-
-Future<Uint8List?> _captureWidgetViewportSlicesPngBytes(
-  OverlayState overlay,
-  Widget Function() buildContent, {
-  required ThemeData theme,
-  required double width,
-  required double pixelRatio,
-  required Size contentSize,
-}) async {
-  final sliceLogicalHeight = _exportCaptureSliceLogicalHeight(
-    pixelRatio: pixelRatio,
-  );
-  final outputWidth = (contentSize.width * pixelRatio).ceil();
-  final outputHeight = (contentSize.height * pixelRatio).ceil();
-  final slices = <({Uint8List bytes, int y})>[];
-
-  double top = 0;
-  while (top < contentSize.height) {
-    final height = (contentSize.height - top).clamp(0.0, sliceLogicalHeight);
-    final boundaryKey = GlobalKey();
-    late OverlayEntry entry;
-    entry = OverlayEntry(
-      builder: (ctx) {
-        return Positioned(
-          left: -10000,
-          top: -10000,
-          child: _ExportCaptureViewportRoot(
-            theme: theme,
-            boundaryKey: boundaryKey,
-            width: width,
-            viewportHeight: height,
-            contentHeight: contentSize.height,
-            offsetY: top,
-            child: buildContent(),
-          ),
-        );
-      },
-    );
-
-    overlay.insert(entry);
-    try {
-      await _waitForExportCaptureFrames(
-        frameCount: 2,
-        settleDelay: const Duration(milliseconds: 80),
-      );
-      final boundary =
-          boundaryKey.currentContext?.findRenderObject()
-              as RenderRepaintBoundary?;
-      if (boundary == null) return null;
-      final slice = await _captureFullBoundaryPngBytes(
-        boundary,
-        pixelRatio: pixelRatio,
-      );
-      if (slice == null) return null;
-      slices.add((bytes: slice, y: (top * pixelRatio).round()));
-    } finally {
-      entry.remove();
-    }
-
-    top += height;
-    await WidgetsBinding.instance.endOfFrame;
-  }
-
-  return _processCapturedExportPng(
-    _CapturedExportPngProcessingRequest.slices(
-      outputWidth: outputWidth,
-      outputHeight: outputHeight,
-      slices: slices
-          .map(
-            (slice) => _ExportPngSlicePayload(bytes: slice.bytes, y: slice.y),
-          )
-          .toList(growable: false),
-      preservePadding: _exportImageBlankTrimPreservePaddingPhysical,
-    ),
-  );
-}
-
-Future<void> _waitForExportCaptureFrames({
-  required int frameCount,
-  Duration settleDelay = Duration.zero,
-}) async {
-  for (var i = 0; i < frameCount; i += 1) {
-    await WidgetsBinding.instance.endOfFrame;
-  }
-  if (settleDelay > Duration.zero) {
-    await Future<void>.delayed(settleDelay);
-  }
-}
-
-bool _shouldUseFullExportCapture(
-  Size logicalSize, {
-  required double pixelRatio,
-}) {
-  return _exportFullCapturePixelRatio(
-        logicalSize,
-        requestedPixelRatio: pixelRatio,
-      ) !=
-      null;
-}
-
-double? _exportFullCapturePixelRatio(
-  Size logicalSize, {
-  required double requestedPixelRatio,
-}) {
-  if (logicalSize.isEmpty || requestedPixelRatio <= 0) return null;
-  final maxLogicalDimension = math.max(logicalSize.width, logicalSize.height);
-  if (maxLogicalDimension <= 0) return null;
-  final requestedPhysicalDimension = maxLogicalDimension * requestedPixelRatio;
-  if (requestedPhysicalDimension <= _maxExportFullCapturePhysicalDimension) {
-    return requestedPixelRatio;
-  }
-  final cappedPixelRatio =
-      _maxExportFullCapturePhysicalDimension / maxLogicalDimension;
-  if (cappedPixelRatio < _minExportFullCapturePixelRatio) return null;
-  return math.min(requestedPixelRatio, cappedPixelRatio);
-}
-
-double _exportCaptureSliceLogicalHeight({required double pixelRatio}) {
-  final height = (_maxExportCaptureSlicePhysicalHeight / pixelRatio).floor();
-  return math.max(height, 1).toDouble();
-}
-
-Future<Uint8List?> _captureFullBoundaryPngBytes(
-  RenderRepaintBoundary boundary, {
-  required double pixelRatio,
-}) async {
-  ui.Image? image;
-  try {
-    image = await boundary.toImage(pixelRatio: pixelRatio);
-    final expectedWidth = (boundary.size.width * pixelRatio).ceil();
-    final expectedHeight = (boundary.size.height * pixelRatio).ceil();
-    if (image.width < expectedWidth || image.height < expectedHeight) {
-      debugPrint(
-        'Full export image capture was clipped '
-        '(${image.width}x${image.height}, expected '
-        '${expectedWidth}x$expectedHeight); falling back to slices.',
-      );
-      return null;
-    }
-    final data = await image.toByteData(
-      format: ui.ImageByteFormat.rawStraightRgba,
-    );
-    if (data == null) return null;
-    // 8-bit straight-alpha readback: ImageByteFormat.png on wide-gamut
-    // backends (iOS Impeller) embeds 10/16-bit wide-gamut bytes that
-    // downstream consumers misinterpret as sRGB. Encode to a plain
-    // 8-bit PNG here instead.
-    return image_lib.encodePng(
-      image_lib.Image.fromBytes(
-        width: image.width,
-        height: image.height,
-        bytes: data.buffer,
-        numChannels: 4,
-      ),
-    );
-  } catch (e) {
-    debugPrint('Full export image capture failed, falling back to slices: $e');
-    return null;
-  } finally {
-    image?.dispose();
-  }
-}
-
-Future<Uint8List> _processCapturedExportPng(
-  _CapturedExportPngProcessingRequest request,
-) {
-  return compute(
-    _processCapturedExportPngSync,
-    request,
-    debugLabel: 'export-image-png-processing',
-  );
-}
-
-@visibleForTesting
-Future<Uint8List> processCapturedExportPngForTesting({
-  Uint8List? singlePngBytes,
-  int? outputWidth,
-  int? outputHeight,
-  List<({Uint8List bytes, int y})>? slices,
-  required int preservePadding,
-}) {
-  final request = singlePngBytes != null
-      ? _CapturedExportPngProcessingRequest.single(
-          singlePngBytes: singlePngBytes,
-          preservePadding: preservePadding,
-        )
-      : _CapturedExportPngProcessingRequest.slices(
-          outputWidth: outputWidth ?? 0,
-          outputHeight: outputHeight ?? 0,
-          slices: (slices ?? const <({Uint8List bytes, int y})>[])
-              .map(
-                (slice) =>
-                    _ExportPngSlicePayload(bytes: slice.bytes, y: slice.y),
-              )
-              .toList(growable: false),
-          preservePadding: preservePadding,
-        );
-  return _processCapturedExportPng(request);
-}
-
-Uint8List _processCapturedExportPngSync(
-  _CapturedExportPngProcessingRequest request,
-) {
-  final singlePngBytes = request.singlePngBytes;
-  if (singlePngBytes != null) {
-    return _trimExportPngBlankPadding(
-      singlePngBytes,
-      preservePadding: request.preservePadding,
-    );
-  }
-
-  final stitched = _stitchExportPngSlicesImage(
-    outputWidth: request.outputWidth,
-    outputHeight: request.outputHeight,
-    slices: request.slices
-        .map((slice) => (bytes: slice.bytes, y: slice.y))
-        .toList(growable: false),
-  );
-  return _encodeTrimmedExportImage(
-    stitched,
-    preservePadding: request.preservePadding,
-  );
-}
-
-class _CapturedExportPngProcessingRequest {
-  const _CapturedExportPngProcessingRequest.single({
-    required this.singlePngBytes,
-    required this.preservePadding,
-  }) : outputWidth = 0,
-       outputHeight = 0,
-       slices = const <_ExportPngSlicePayload>[];
-
-  const _CapturedExportPngProcessingRequest.slices({
-    required this.outputWidth,
-    required this.outputHeight,
-    required this.slices,
-    required this.preservePadding,
-  }) : singlePngBytes = null;
-
-  final Uint8List? singlePngBytes;
-  final int outputWidth;
-  final int outputHeight;
-  final List<_ExportPngSlicePayload> slices;
-  final int preservePadding;
-}
-
-class _ExportPngSlicePayload {
-  const _ExportPngSlicePayload({required this.bytes, required this.y});
-
-  final Uint8List bytes;
-  final int y;
-}
-
-Uint8List _stitchExportPngSlices({
-  required int outputWidth,
-  required int outputHeight,
-  required List<({Uint8List bytes, int y})> slices,
-}) {
-  return image_lib.encodePng(
-    _stitchExportPngSlicesImage(
-      outputWidth: outputWidth,
-      outputHeight: outputHeight,
-      slices: slices,
-    ),
-  );
-}
-
-image_lib.Image _stitchExportPngSlicesImage({
-  required int outputWidth,
-  required int outputHeight,
-  required List<({Uint8List bytes, int y})> slices,
-}) {
-  final stitched = image_lib.Image(
-    width: outputWidth,
-    height: outputHeight,
-    numChannels: 4,
-  )..clear(image_lib.ColorRgba8(0, 0, 0, 0));
-
-  for (final slice in slices) {
-    final decoded = image_lib.decodePng(slice.bytes);
-    if (decoded == null) continue;
-    final dstY = math.max(slice.y, 0);
-    final srcY = math.max(-slice.y, 0);
-    final copyHeight = math.min(decoded.height - srcY, outputHeight - dstY);
-    final copyWidth = math.min(decoded.width, outputWidth);
-    if (copyWidth <= 0 || copyHeight <= 0) continue;
-
-    for (var y = 0; y < copyHeight; y += 1) {
-      for (var x = 0; x < copyWidth; x += 1) {
-        stitched.setPixel(x, dstY + y, decoded.getPixel(x, srcY + y));
-      }
-    }
-  }
-
-  return stitched;
-}
-
-@visibleForTesting
-Uint8List trimExportPngBlankPaddingForTesting(
-  Uint8List bytes, {
-  required int preservePadding,
-}) {
-  return _trimExportPngBlankPadding(bytes, preservePadding: preservePadding);
-}
-
-Uint8List _trimExportPngBlankPadding(
-  Uint8List bytes, {
-  required int preservePadding,
-}) {
-  final decoded = image_lib.decodePng(bytes);
-  if (decoded == null || decoded.width <= 0 || decoded.height <= 0) {
-    return bytes;
-  }
-  return _encodeTrimmedExportImage(decoded, preservePadding: preservePadding);
-}
-
-Uint8List _encodeTrimmedExportImage(
-  image_lib.Image decoded, {
-  required int preservePadding,
-}) {
-  final trimmed = _trimExportImageBlankPadding(
-    decoded,
-    preservePadding: preservePadding,
-  );
-  return image_lib.encodePng(trimmed);
-}
-
-image_lib.Image _trimExportImageBlankPadding(
-  image_lib.Image decoded, {
-  required int preservePadding,
-}) {
-  final background = _exportImageBlankReferencePixel(decoded);
-  final width = decoded.width;
-  final height = decoded.height;
-
-  int? top;
-  for (var y = 0; y < height; y += 1) {
-    if (_exportImageRowHasContent(decoded, y, background)) {
-      top = y;
-      break;
-    }
-  }
-  if (top == null) return decoded;
-
-  var bottom = height - 1;
-  for (var y = height - 1; y >= top; y -= 1) {
-    if (_exportImageRowHasContent(decoded, y, background)) {
-      bottom = y;
-      break;
-    }
-  }
-
-  var left = 0;
-  for (var x = 0; x < width; x += 1) {
-    if (_exportImageColumnHasContent(decoded, x, top, bottom, background)) {
-      left = x;
-      break;
-    }
-  }
-
-  var right = width - 1;
-  for (var x = width - 1; x >= left; x -= 1) {
-    if (_exportImageColumnHasContent(decoded, x, top, bottom, background)) {
-      right = x;
-      break;
-    }
-  }
-
-  final padding = math.max(preservePadding, 0);
-  final cropLeft = math.max(left - padding, 0);
-  final cropTop = math.max(top - padding, 0);
-  final cropRight = math.min(right + padding, width - 1);
-  final cropBottom = math.min(bottom + padding, height - 1);
-
-  if (cropLeft == 0 &&
-      cropTop == 0 &&
-      cropRight == width - 1 &&
-      cropBottom == height - 1) {
-    return decoded;
-  }
-
-  return image_lib.copyCrop(
-    decoded,
-    x: cropLeft,
-    y: cropTop,
-    width: cropRight - cropLeft + 1,
-    height: cropBottom - cropTop + 1,
-  );
-}
-
-image_lib.Pixel _exportImageBlankReferencePixel(image_lib.Image image) {
-  final corners = [
-    image.getPixel(0, 0),
-    image.getPixel(image.width - 1, 0),
-    image.getPixel(0, image.height - 1),
-    image.getPixel(image.width - 1, image.height - 1),
-  ];
-  for (final pixel in corners) {
-    if (pixel.a <= _exportImageBlankAlphaTolerance) return pixel;
-  }
-  return corners.first;
-}
-
-bool _exportImageRowHasContent(
-  image_lib.Image image,
-  int y,
-  image_lib.Pixel background,
-) {
-  for (var x = 0; x < image.width; x += 1) {
-    if (!_exportImageIsBlankPixel(image.getPixel(x, y), background)) {
-      return true;
-    }
-  }
-  return false;
-}
-
-bool _exportImageColumnHasContent(
-  image_lib.Image image,
-  int x,
-  int top,
-  int bottom,
-  image_lib.Pixel background,
-) {
-  for (var y = top; y <= bottom; y += 1) {
-    if (!_exportImageIsBlankPixel(image.getPixel(x, y), background)) {
-      return true;
-    }
-  }
-  return false;
-}
-
-bool _exportImageIsBlankPixel(
-  image_lib.Pixel pixel,
-  image_lib.Pixel background,
-) {
-  if (pixel.a <= _exportImageBlankAlphaTolerance) return true;
-  if (background.a <= _exportImageBlankAlphaTolerance) return false;
-  return _exportImageChannelNear(pixel.r, background.r) &&
-      _exportImageChannelNear(pixel.g, background.g) &&
-      _exportImageChannelNear(pixel.b, background.b) &&
-      _exportImageChannelNear(pixel.a, background.a);
-}
-
-bool _exportImageChannelNear(num a, num b) {
-  return (a - b).abs() <= _exportImageBlankColorTolerance;
 }
 
 Future<void> showMessageExportSheet(
